refactor: decompose Dashboard.tsx into sub-components
Extract presentational components from the 1,821-line Dashboard.tsx:
- dashboard-types.ts: shared interfaces (Account, Transaction, etc.)
- TimeRangeSelector.tsx: time range buttons + custom date pickers
- AccountFilterPopover.tsx: account filter popover with checkboxes
- SummaryCards.tsx: net worth, accounts, transactions summary cards
- NetWorthChart.tsx: line chart for net worth evolution
- AssetAllocationChart.tsx: doughnut chart for asset allocation
- ExpensesByCategoryChart.tsx: doughnut chart for expenses by category
- IncomeVsExpensesChart.tsx: bar chart for income vs expenses

Parent Dashboard.tsx retains all state, data fetching, and useMemo
computations. Child components are purely presentational, receiving
computed chart data and options as props.
