+++ app/src/features/dashboard/NetWorthChart.tsx
@@ -1,0 +1,42 @@
+import type { ChartOptions, ChartData } from "chart.js";
+import { Line } from "react-chartjs-2";
+import { useTranslation } from "react-i18next";
+
+interface NetWorthChartProps {
+  chartData: ChartData<"line"> | null;
+  options: ChartOptions<"line">;
+}
+
+export default function NetWorthChart({
+  chartData,
+  options,
+}: NetWorthChartProps) {
+  const { t } = useTranslation();
+
+  return (
+    <div className="chart-container">
+      <div className="chart-header">
+        <h3 className="chart-title">{t("dashboard.networth_evolution")}</h3>
+        <p className="chart-subtitle">
+          {t("dashboard.subtitle.networth_growth")}
+        </p>
+      </div>
+      <div className="chart-wrapper">
+        <div className="chart-body">
+          {chartData ? (
+            <Line options={options} data={chartData} />
+          ) : (
+            <div className="loading-container">
+              <div className="loading-content">
+                <div className="loading-spinner"></div>
+                <span className="loading-text">
+                  {t("loading.loading_data")}
+                </span>
+              </div>
+            </div>
+          )}
+        </div>
+      </div>
+    </div>
+  );
+}
